src/test/java/it/polimi/ingsw/Model/Cards/Constraints/AdjacentRoomTest.java
package it.polimi.ingsw.Model.Cards.Constraints;

import it.polimi.ingsw.Exception.NotFoundException;
import it.polimi.ingsw.Model.Map.Board;
import it.polimi.ingsw.Model.Player;
import it.polimi.ingsw.Model.TargetParameter;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class AdjacentRoomTest {
    AdjacentRoom test;
    Player owner;
    Board board;
    TargetParameter target;

    @BeforeEach
    public void setup(){
        board = new Board(null,"./resources/Board/Board1.json");
        owner = new Player(true,"Yellow","Bruno");
        test = new AdjacentRoom();
        target = new TargetParameter(null,owner,null,null,null);
    }

    @Test
    void nearRoom(){
        try {
            owner.setPosition(board.find(1,2));
        } catch (NotFoundException e) {
            e.printStackTrace();
        }
        target.setTargetRoom(board.getRooms().get(0));
        assertEquals(true,test.canShoot(target));
    }

    @Test
    void notNearRoom(){
        try {
            owner.setPosition(board.find(2,2));
        } catch (NotFoundException e) {
            e.printStackTrace();
        }
        target.setTargetRoom(board.getRooms().get(0));
        assertEquals(false,test.canShoot(target));
    }

    @Test
    void inMyRoom(){
        try {
            owner.setPosition(board.find(1,1));
        } catch (NotFoundException e) {
            e.printStackTrace();
        }
        target.setTargetRoom(board.getRooms().get(0));
        assertEquals(false,test.canShoot(target));
    }
}